Fix en marcar favorito

+++ src/main/java/com/utn/tacs/tp2016c1g4/marvel_webapp/api/request/favorito/FavoritoPutRequest.java
@@ -1,21 +1,21 @@
 package com.utn.tacs.tp2016c1g4.marvel_webapp.api.request.favorito;
 
 public class FavoritoPutRequest {
-	
-	private Integer idPersonaje;
-	
+
+	private Long idPersonaje;
+
 	public FavoritoPutRequest() {
 	}
 
-	public FavoritoPutRequest(Integer idPersonaje) {
+	public FavoritoPutRequest(Long idPersonaje) {
 		this.setIdPersonaje(idPersonaje);
 	}
 
-	public Integer getIdPersonaje() {
+	public Long getIdPersonaje() {
 		return idPersonaje;
 	}
 
-	public void setIdPersonaje(Integer idPersonaje) {
+	public void setIdPersonaje(Long idPersonaje) {
 		this.idPersonaje = idPersonaje;
 	}
 }
